Sources/ETNM.PlanMe/ETNM.PlanMe/Controllers/HomeController.cs
using System.Web.Mvc;

namespace ETNM.PlanMe.Controllers
{
    public class HomeController : Controller
    {
        public ActionResult Index()
        {
            ViewBag.Message = "Welcome to the PlanMe application.";

            return View();
        }

        public ActionResult About()
        {
            ViewBag.Message = "Your app description page.";

            return View();
        }

        public ActionResult Contact()
        {
            ViewBag.Message = "Your contact page.";

            return View();
        }
    }
}
